+++ packages/camera/camera_android_camerax/android/src/main/java/io/flutter/plugins/camerax/DeviceOrientationManager.java
@@ -10,8 +10,10 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.res.Configuration;
 import android.view.Display;
+import android.view.OrientationEventListener;
 import android.view.Surface;
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
 import io.flutter.embedding.engine.systemchannels.PlatformChannel;
 import io.flutter.embedding.engine.systemchannels.PlatformChannel.DeviceOrientation;
@@ -28,6 +30,8 @@ public class DeviceOrientationManager {
   private PlatformChannel.DeviceOrientation lastOrientation;
   private BroadcastReceiver broadcastReceiver;
 
+  @VisibleForTesting @Nullable protected OrientationEventListener orientationEventListener;
+
   DeviceOrientationManager(DeviceOrientationManagerProxyApi api) {
     this.api = api;
   }
@@ -38,39 +42,46 @@ public class DeviceOrientationManager {
   }
 
   /**
-   * Starts listening to the device's sensors or UI for orientation updates.
+   * Starts listening to the device's sensors for device orientation updates.
    *
    * <p>When orientation information is updated, the callback method of the {@link
    * DeviceOrientationManagerProxyApi} is called with the new orientation.
-   *
-   * <p>If the device's ACCELEROMETER_ROTATION setting is enabled the {@link
-   * DeviceOrientationManager} will report orientation updates based on the sensor information. If
-   * the ACCELEROMETER_ROTATION is disabled the {@link DeviceOrientationManager} will fallback to
-   * the deliver orientation updates based on the UI orientation.
    */
   public void start() {
     stop();
 
-    broadcastReceiver =
-        new BroadcastReceiver() {
-          @Override
-          public void onReceive(Context context, Intent intent) {
-            handleUIOrientationChange();
-          }
-        };
-    getContext().registerReceiver(broadcastReceiver, orientationIntentFilter);
-    broadcastReceiver.onReceive(getContext(), null);
+    // Listen for changes in device orientation at the default rate that is suitable for monitoring
+    // typical screen orientation changes.
+    orientationEventListener = createOrientationEventListener();
+    orientationEventListener.enable();
+  }
+
+  @VisibleForTesting
+  @NonNull
+  /**
+   * Creates an {@link OrientationEventListener} that will call the callback method of the {@link
+   * DeviceOrientationManagerProxyApi} whenever it is notified of a new device orientation and this
+   * {@code DeviceOrientationManager} instance determines that the orientation of the device {@link
+   * Configuration} has changed.
+   */
+  protected OrientationEventListener createOrientationEventListener() {
+    return new OrientationEventListener(getContext()) {
+      @Override
+      public void onOrientationChanged(int orientation) {
+        handleUIOrientationChange();
+      }
+    };
   }
 
   /** Stops listening for orientation updates. */
   public void stop() {
-    if (broadcastReceiver == null) {
+    if (orientationEventListener == null) {
       return;
     }
-    getContext().unregisterReceiver(broadcastReceiver);
-    broadcastReceiver = null;
-
     lastOrientation = null;
+
+    orientationEventListener.disable();
+    orientationEventListener = null;
   }
 
   /**
@@ -87,8 +98,7 @@ public class DeviceOrientationManager {
   }
 
   /**
-   * Handles orientation changes coming from either the device's sensors or the
-   * OrientationIntentFilter.
+   * Handles orientation changes coming from the device's sensors.
    *
    * <p>This method is visible for testing purposes only and should never be used outside this
    * class.